EBR-3: Add final feature

+++ ebook-reader/src/main/java/com/aritan/ebook_reader/features/category/CategoryService.java
@@ -2,13 +2,16 @@ package com.aritan.ebook_reader.features.category;
 
 import com.aritan.ebook_reader.common.constants.messages.book.CategoryMessage;
 import com.aritan.ebook_reader.common.exception.ResourceNotFoundException;
+import com.aritan.ebook_reader.common.exception.ValidationException;
 import com.aritan.ebook_reader.common.models.book.Category;
+import com.aritan.ebook_reader.features.book.IBookRepository;
 import com.aritan.ebook_reader.features.category.dtos.CategoryCreateRequest;
 import com.aritan.ebook_reader.features.category.dtos.CategoryFilterRequest;
 import com.aritan.ebook_reader.features.category.dtos.CategoryResponse;
 import com.aritan.ebook_reader.features.category.dtos.CategoryUpdateRequest;
 import com.aritan.ebook_reader.features.category.utilities.CategoryMapper;
 import com.aritan.ebook_reader.features.category.utilities.CategorySpecification;
+import com.aritan.ebook_reader.features.category.utilities.CategoryValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -16,13 +19,15 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
+import java.util.*;
 
 @Service
 @RequiredArgsConstructor
 public class CategoryService implements ICategoryService{
     private final ICategoryRepository categoryRepository;
+    private final IBookRepository bookRepository;
     private final CategoryMapper categoryMapper;
+    private final CategoryValidator validator;
     @Override
     public Page<CategoryResponse> getAllCategoriesByAdmin(CategoryFilterRequest request, Pageable page) {
         Specification<Category> spec = Specification
@@ -42,7 +47,12 @@ public class CategoryService implements ICategoryService{
     }
 
     @Override
+    @Transactional
     public List<CategoryResponse> createCategory(List<CategoryCreateRequest> requests) {
+        validator.validateNoDuplicateInBatch(requests);
+        requests.forEach(r ->
+                validator.validateUniqueForCreate(r.getCategoryName(), r.getSlug()));
+
         List<Category> categories = requests.stream()
                         .map(categoryMapper::toCategory)
                         .toList();
@@ -60,6 +70,11 @@ public class CategoryService implements ICategoryService{
                         String.format(CategoryMessage.CATEGORY_NOT_FOUND, categoryId)
                 ));
 
+        validator.validateUniqueForUpdate(
+                categoryId,
+                updateRequest.getCategoryName(),
+                updateRequest.getSlug());
+
         categoryMapper.updateCategory(updateRequest, existedCategory);
         Category updatedCategory = categoryRepository.save(existedCategory);
 
@@ -67,11 +82,18 @@ public class CategoryService implements ICategoryService{
     }
 
     @Override
+    @Transactional
     public void deleteCategory(Long categoryId) {
         Category category = categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new ResourceNotFoundException(
                         String.format(CategoryMessage.CATEGORY_NOT_FOUND, categoryId)
                 ));
+
+        if (bookRepository.existsByCategories_CategoryId(categoryId)) {
+            throw new ValidationException(
+                    String.format(CategoryMessage.IS_HAS_BOOKS_ASSIGNED, categoryId)
+            );
+        }
 
         categoryRepository.delete(category);
     }
